move the login handler into an auth folder
